# Improve Chunk.empty concatenation 50x for 1.x