Inspired by https://github.com/zio/zio/pull/7396 .

The nonempty case is a little slower. The empty case is a lot faster. I'm unsure if the tradeoff is worth it, but I had this idea and wanted to present it.

Before
```
[info] Benchmark                                      (size)   Mode  Cnt          Score         Error  Units
[info] ChunkConcatBenchmarks.balancedConcatOnce        10000  thrpt   10  203992533.765 ± 2075468.906  ops/s
[info] ChunkConcatBenchmarks.balancedConcatRecursive   10000  thrpt   10      18370.965 ±      41.063  ops/s
[info] ChunkConcatBenchmarks.leftConcat0               10000  thrpt   10       1509.055 ±       8.703  ops/s
[info] ChunkConcatBenchmarks.leftConcat1               10000  thrpt   10       1516.642 ±      13.518  ops/s
[info] ChunkConcatBenchmarks.leftConcat10              10000  thrpt   10      21650.274 ±     162.907  ops/s
[info] ChunkConcatBenchmarks.rightConcat0              10000  thrpt   10       1456.520 ±      12.798  ops/s
[info] ChunkConcatBenchmarks.rightConcat1              10000  thrpt   10       1497.503 ±      15.500  ops/s
[info] ChunkConcatBenchmarks.rightConcat10             10000  thrpt   10      20575.558 ±     203.644  ops/s
```
After
```
[info] Benchmark                                      (size)   Mode  Cnt          Score         Error  Units
[info] ChunkConcatBenchmarks.balancedConcatOnce        10000  thrpt   10  206777416.179 ± 1702134.374  ops/s
[info] ChunkConcatBenchmarks.balancedConcatRecursive   10000  thrpt   10      17183.683 ±     203.886  ops/s
[info] ChunkConcatBenchmarks.leftConcat0               10000  thrpt   10      75478.279 ±     542.242  ops/s
[info] ChunkConcatBenchmarks.leftConcat1               10000  thrpt   10       1373.804 ±      96.077  ops/s
[info] ChunkConcatBenchmarks.leftConcat10              10000  thrpt   10      19332.971 ±     149.147  ops/s
[info] ChunkConcatBenchmarks.rightConcat0              10000  thrpt   10      67146.335 ±     509.862  ops/s
[info] ChunkConcatBenchmarks.rightConcat1              10000  thrpt   10       1336.395 ±      10.295  ops/s
[info] ChunkConcatBenchmarks.rightConcat10             10000  thrpt   10      19059.703 ±     131.697  ops/s
```